+++ package.json
@@ -11,14 +11,15 @@
   },
   "dependencies": {
     "@types/react-modal": "^3.16.3",
-    "react-modal": "^3.16.1",
     "axios": "^1.6.1",
     "firebase": "^10.5.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.11.0",
+    "react-modal": "^3.16.1",
     "react-router-dom": "^6.18.0",
     "react-spinners": "^0.13.8",
+    "socket.io-client": "^4.7.2",
     "styled-components": "^6.1.0",
     "styled-reset": "^4.5.1"
   },
